requirements.txt
pyTelegramBotAPI>=4.29.0
python-dotenv>=1.0.0
pdfplumber>=0.10.0
PyPDF2>=3.0.0

